feat: authentication using HTTP Basic
This commit restricts writing to repositories to users having a JWT with
the correct namespace claim. For example, to write to a repository
"humanperson/somerepo", the JWT token must include the claim
{ "namespace": [ "humanperson" ] } or
{ "namespace": [ "humanperson/somerepo" ] }.

It works with the Negotiate protocol used by Git LFS, so no further
configuration is needed on the client side. They will receive a dialog
box asking for username and password, where they can fill anything in
the username field and token in the password field.